migrations for board and card models
