+++ app/models/card.py
@@ -1,14 +1,14 @@
 from app import db
 
 class Card(db.Model):
-    card_id = db.Column(db.Integer, primary_key=True, autouncrement=True)
+    card_id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     message = db.Column(db.String(225), nullable=False)
     likes_count= db.Column(db.Integer, nullable=False )
 
 def to_dict(self):
     return ({"card_id": self.card_id,
-             "message": self.message,
-             "likes_count": self.likes_count})
+            "message": self.message,
+            "likes_count": self.likes_count})
 
 @classmethod
 def from_dict(cls,data_dict):
